Add simple PBAC access control, cache, json patch
Replaces role-based board member management with a permission-based system. Updates frontend and backend to use JSON Patch for partial updates to boards, lists, and cards. Removes role selection from member management UI, adds permission management popover, and updates API endpoints and DTOs accordingly. Cleans up error handling, improves optimistic UI updates for drag-and-drop, and updates Keycloak config for longer token lifespan and PKCE support.

+++ src/Application/DTOs/ListCards/ListCardResponse.cs
@@ -8,5 +8,5 @@ public sealed record ListCardResponse
     public string Title { get; set; } = string.Empty;
     public string? Description { get; set; } = string.Empty;
     public DateTime? CompletedAt { get; set; }
-    public int Position { get; set; }
+    public double Position { get; set; }
 }
